requirements.txt
aiohttp==3.8.3 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "4.0"
bilireq==0.2.3.post0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows" or python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
commonmark==0.9.1 ; python_version >= "3.9" and python_version < "4.0"
creart-graia==0.1.5 ; python_version >= "3.9" and python_version < "4.0"
creart==0.2.2 ; python_version >= "3.9" and python_version < "4.0"
croniter==1.3.8 ; python_version >= "3.9" and python_version < "4.0"
cryptography==39.0.0 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.88.0 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.9" and python_version < "4.0"
graia-amnesia==0.7.1 ; python_version >= "3.9" and python_version < "4.0"
graia-ariadne[standard]==0.10.3 ; python_version >= "3.9" and python_version < "4.0"
graia-broadcast==0.19.1 ; python_version >= "3.9" and python_version < "4.0"
graia-saya==0.0.17 ; python_version >= "3.9" and python_version < "4.0"
graia-scheduler==0.1.4 ; python_version >= "3.9" and python_version < "4.0"
graiax-playwright==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.51.1 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.23.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.0.0 ; python_version >= "3.9" and python_version < "4.0"
launart==0.6.2 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
peewee==3.15.4 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.9" and python_version < "4.0"
playwright==1.29.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.21.12 ; python_version >= "3.9" and python_version < "4.0"
psutil==5.9.4 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.9" and python_version < "4.0"
pyee==9.0.4 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
qrcode==7.3.1 ; python_version >= "3.9" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
rich==13.0.0 ; python_version >= "3.9" and python_version < "4.0"
richuru==0.1.1 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
sentry-sdk==1.12.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
statv==0.3.2 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.13 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
yarl==1.8.2 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.11.0 ; python_version >= "3.9" and python_version < "4.0"
